chore(intent): relocate queryKeys.ts from backends/occt/ to intent/
The headless build boundary spec required CaptureSession (and any
agent-facing layer) to not import OCCT backend modules directly.
EDGE_QUERY_KEYS / FACE_QUERY_KEYS are vocabulary constants (the legal
keys of EdgeQuery / FaceQuery) — vocabulary belongs in src/intent/, not
src/backends/occt/.

The EdgeQuery / FaceQuery TYPES still live with the OCCT implementation
(they describe an OCCT-backed query DSL); this slice moves only the
key arrays. Three importers updated; test file moved to mirror the new
location.
