+++ tests/unit/intent/queryKeys.test.ts
@@ -1,6 +1,6 @@
-// tests/unit/backends/occt/queryKeys.test.ts
+// tests/unit/intent/queryKeys.test.ts
 import { describe, it, expect } from 'vitest';
-import { EDGE_QUERY_KEYS, FACE_QUERY_KEYS } from '../../../../src/backends/occt/queryKeys';
+import { EDGE_QUERY_KEYS, FACE_QUERY_KEYS } from '../../../src/intent/queryKeys';
 
 describe('queryKeys (single source of truth)', () => {
   it('EDGE_QUERY_KEYS is the canonical key list (compile-time exhaustiveness in queryKeys.ts)', () => {
@@ -26,9 +26,9 @@ describe('queryKeys (single source of truth)', () => {
   it('all consumers import from the same source (modules load cleanly)', async () => {
     // The compile-time `keyof` checks above are the actual drift guard.
     // This test just ensures all 3 consumer files load without import errors.
-    const fromCaptureSession = await import('../../../../src/capture/captureSession');
-    const fromEdgeSelection = await import('../../../../src/backends/occt/edgeSelection');
-    const fromListApi = await import('../../../../src/mcp/tools/listApi');
+    const fromCaptureSession = await import('../../../src/capture/captureSession');
+    const fromEdgeSelection = await import('../../../src/backends/occt/edgeSelection');
+    const fromListApi = await import('../../../src/mcp/tools/listApi');
     expect(fromCaptureSession).toBeDefined();
     expect(fromEdgeSelection).toBeDefined();
     expect(fromListApi).toBeDefined();
